pyproject.toml
[project]
name = "merlin-vlm"
version = "0.0.6"
requires-python = ">=3.9"
description = "Merlin: A Vision Language Foundation Model for 3D Computed Tomography"
readme = "README.md"
license = { file = "LICENSE" }
authors = [
    { name = "Louis Blankemeier", email = "louis.blankemeier@stanford.edu"},
    { name = "Ashwin Kumar", email = "akkumar@stanford.edu"},
    { name = "Stanford Machine Intelligence for Medical Imaging (MIMI)" }
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Healthcare Industry",
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
]
dependencies = [
    "numpy>=1.26.4", # tested with 1.26.4
    "torch>=2.1.2", # tested with 2.1.2
    "monai>=1.3.0", # tested with 1.3.0
    "huggingface_hub", # tested with 0.30.1
    "transformers>=4.38.2", # tested with 4.38.2
    "torchvision>=0.16.2", # tested with 0.16.2
    "nibabel", # tested with 5.3.2
    "nltk", # tested with 3.9.1
    "pandas", # tested with 2.3.2
    "rich", # tested with 14.1.0
    "einops", # tested with 0.8.1
    "peft", # tested with 0.10.0
    "matplotlib", # tested with 3.9.4
    "sentencepiece", # tested with 0.2.1
    "protobuf", # tested with 6.32.1,
    "accelerate", # tested with 0.34.2
]

[project.urls]
homepage = "https://github.com/StanfordMIMI/Merlin"
repository = "https://github.com/StanfordMIMI/Merlin"

[project.optional-dependencies]
dev = [
    "ruff",
    "pre-commit",
    "mdformat"
]
gguf = [
    "llama-cpp-python>=0.2.0",
]
eval = [
    "rouge_score",
    "bert_score",
]

[build-system]
requires = ["setuptools>=67.8.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["merlin", "merlin.data", "merlin.models", "merlin.utils"]
exclude = ["documentation"]

[tool.codespell]
skip = '.git,*.pdf,*.svg, *.png'